+++ derive-input/src/internal/dsl/method/get_many_options.rs
