+++ derive-output/src/output.rs
