--- derive/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-authors = ["Tamaro Skaljic<mail@tamaro-skaljic.de>"]
-name = "spacetimedsl_derive"
-version = "0.4.0"
-edition = "2024"
-description = "Macros to extend SpacetimeDSL. You shouldn't depend on this directly and instead use the spacetimedsl crate."
-documentation = "https://docs.rs/spacetimedsl_derive"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/tamaro-skaljic/SpacetimeDSL/derive"
-publish = true
-include = ["/src/**"]
-keywords = ["dsl", "spacetime", "macro", "codegen", "boilerplate"]
-
-[lib]
-proc-macro = true
-
-[dependencies]
-# https://crates.io/crates/ident_case Utility for applying case rules to Rust identifiers.
-ident_case = "1.0.1"
-
-# https://crates.io/crates/proc_macro2 A substitute implementation of the compiler's `proc_macro` API to decouple token-based libraries from the procedural macro use case.
-proc-macro2 = "1.0"
-
-# https://crates.io/crates/syn Parser for Rust source code
-syn = { version = "2.0" }
-
-# https://crates.io/crates/quote Quasi-quoting macro quote!(...)
-quote = "1.0"
-
-# https://crates.io/crates/spacetimedsl_derive-output Contains structs which represent the output of SpacetimeDSL. Can be used by other proc-macro crates to build Macros which utilize SpacetimeDSL under the hood.
-spacetimedsl_derive-output = { version = "0.4.0", path = "../derive-output" }
