+++ publish-lib_derive-input.cmd
@@ -1,0 +1,6 @@
+@echo off
+
+echo Publishing lib_derive-input...
+cd derive-input
+cargo publish --allow-dirty
+cd ..
