+++ publish-lib_derive-output.cmd
@@ -1,0 +1,6 @@
+@echo off
+
+echo Publishing lib_derive-output...
+cd derive-output
+cargo publish --allow-dirty
+cd ..
