--- publish-lib_derive.cmd
@@ -1,6 +1,0 @@
-@echo off
-
-echo Publishing lib_derive...
-cd derive
-cargo publish --allow-dirty
-cd ..
